requirements.txt
streamlit==1.29.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
requests-html==0.10.0
httpx
